pri: Fix decoding of GCR sectors that start at or near the index
If a sector starts right after the index but the sync bytes preceding
it are before the index, there might not yet be byte synchronization
and the sector will be missed.

With this patch the track is scanned from the first index past the
second index to the position of the first decoded sector.