ini: Add new directive 'rinclude'
There are now two keywords to include a file that differ
if the specified file name is not absolute:

include: File names are relative to the current working
directory.

rinclude: File names are relative to the parent directory of
the file in which the rinclude keyword appears. This is how
include behaved before this patch.